its_learning_scraper/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
